refactoring: Refactoring of ConfigurationActivity
Moved LineSelectionScreenWidget to a separate file and created a destination for the screen so i can work on creating a drawer like the one in MainActivity to enable to configure also the widget for the stops and not only the lines
