+++ app/src/main/java/org/stypox/tridenta/widget/WidgetConfigurationActivity.kt
@@ -5,30 +5,14 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.pulltorefresh.PullToRefreshBox
-import androidx.compose.material3.pulltorefresh.rememberPullToRefreshState
-import androidx.compose.runtime.Composable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.safeDrawingPadding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import androidx.glance.appwidget.GlanceAppWidgetManager
 import androidx.glance.appwidget.state.updateAppWidgetState
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -39,19 +23,14 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
-import org.stypox.tridenta.R
 import org.stypox.tridenta.db.data.DbLine
-import org.stypox.tridenta.enums.Area
 import org.stypox.tridenta.enums.Direction
 import org.stypox.tridenta.extractor.ROME_ZONE_ID
 import org.stypox.tridenta.log.logError
-import org.stypox.tridenta.ui.lines.AreaChip
-import org.stypox.tridenta.ui.lines.LineItem
-import org.stypox.tridenta.ui.lines.LinesUiState
 import org.stypox.tridenta.ui.lines.LinesViewModel
-import org.stypox.tridenta.ui.lines.SelectAreaDialog
 import org.stypox.tridenta.ui.theme.AppTheme
 import org.stypox.tridenta.widget.actions.WidgetEntryPoint
+import org.stypox.tridenta.widget.ui.LineSelectionScreenWidget
 import java.time.ZonedDateTime
 
 @AndroidEntryPoint
@@ -86,17 +65,24 @@ class WidgetConfigurationActivity :
             val lastReloadWasError by linesViewModel.lastReloadWasError.collectAsState()
 
             AppTheme {
-                LineSelectionScreen(
-                    criticalError = linesUiState.error,
-                    minorError = lastReloadWasError,
-                    loading = linesUiState.loading,
-                    onReload = linesViewModel::onReload,
-                    state = linesUiState,
-                    onLineSelected = { selectedLine ->
-                        saveWidgetConfiguration(selectedLine)
-                    },
-                    setSelectedArea = linesViewModel::setSelectedArea
-                )
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    Box(modifier = Modifier.safeDrawingPadding()) {
+                        LineSelectionScreenWidget(
+                            criticalError = linesUiState.error,
+                            minorError = lastReloadWasError,
+                            loading = linesUiState.loading,
+                            onReload = linesViewModel::onReload,
+                            state = linesUiState,
+                            onLineSelected = { selectedLine ->
+                                saveWidgetConfiguration(selectedLine)
+                            },
+                            setSelectedArea = linesViewModel::setSelectedArea
+                        )
+                    }
+                }
             }
         }
     }
@@ -167,62 +153,6 @@ class WidgetConfigurationActivity :
                 }
                 setResult(RESULT_OK, resultValue)
                 finish()
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun LineSelectionScreen(
-    criticalError: Boolean,
-    minorError: Boolean,
-    loading: Boolean,
-    onReload: () -> Unit,
-    state: LinesUiState,
-    onLineSelected: (DbLine) -> Unit,
-    setSelectedArea: (Area) -> Unit
-) {
-    var showAreaDialog by rememberSaveable { mutableStateOf(false) }
-    if (showAreaDialog) {
-        SelectAreaDialog(
-            selectedArea = state.selectedArea,
-            setSelectedArea = setSelectedArea,
-            onDismiss = { showAreaDialog = false }
-        )
-    }
-    Scaffold(topBar = {
-        TopAppBar(title = {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Text(text = stringResource(R.string.selected_area))
-                AreaChip(
-                    area = state.selectedArea,
-                    onClick = { showAreaDialog = true }
-                )
-            }
-        })
-    }) { paddingValues ->
-        PullToRefreshBox(
-            isRefreshing = loading,
-            state = rememberPullToRefreshState(),
-            onRefresh = onReload,
-            modifier = Modifier
-                .padding(paddingValues)
-                .fillMaxHeight()
-        ) {
-            LazyColumn(
-                modifier = Modifier
-                    .fillMaxWidth()
-            ) {
-                items(state.lines) { line ->
-                    LineItem(
-                        line, true,
-                        modifier = Modifier.clickable { onLineSelected(line) },
-                    )
-                }
             }
         }
     }
